Reduce messages sent by .members command
The current way of sending matches to the channel was a crude solution
resulting in an n+1 which, while not an issue on a small server, could
well be on one with hundreds of role members.

This refactors the command so that all results can be grouped and sent
in one batch at the end of the search.