Save username when member has no nick
If a member leaves, then rejoins after changing their username, it could
be difficult to identify them if they had no nickname defined. This
saves the username any time an update is triggered with no nickname,
e.g. on first join or when a nickname is removed. That way there will
always be a name to reference any time somebody rejoins.