Use config to get channel ID for joins/parts
The mention strings included in `m.Content` (eg `<#1242893479>`) won't
work here as we need the plain ID string to pass to ChannelMessageSend,
so there's a new little helper that trims the junk.

We'll still default to the "first channel", but that too could use some
improvement as there's no guarantee that will be a text channel.