Fix lack of members count when list is too long
Discord has a character limit which causes the message to not get sent
when a list of members pushes the message over that cap. The result is
the "Searching X members..." message is sent, but no further output.

This splits up the second message so that count is always sent
regardless. List will need splitting separately.