Send long lists as multiple messages
Discord has a message limit of 2000 characters, so this "buffers" output
into a separate var which is then checked for length before each new
member name is added to it. If adding the current name and closing
backticks would cause the length to exceed 2000 chars, we first send
whatever output is saved already before resetting it.

Fixes display of roles with so many members that only the count is seen.